# openapi.json
{
  "components" : {
    "schemas" : {
      "AllegroContext" : {
        "properties" : {
          "auctionId" : {
            "type" : "string"
          },
          "auctionName" : {
            "type" : "string"
          },
          "buyerLogin" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "MessageInfo" : {
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "name" : {
            "$ref" : "#/components/schemas/MessageInfo",
            "writeOnly" : true
          },
          "ticket" : {
            "format" : "byte",
            "type" : "string"
          },
          "token" : {
            "$ref" : "#/components/schemas/MessageInfo",
            "writeOnly" : true
          },
          "variables" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "type" : "object"
          }
        },
        "type" : "object"
      },
...
      "ParticipationData" : {
        "properties" : {
          "arrivalDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string"
          },
          "createdDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "firstName" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "gender" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "info" : {
            "type" : "string"
          },
          "lastModifiedBy" : {
            "type" : "string"
          },
          "lastModifiedDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "privacyPolicyAccepted" : {
            "type" : "boolean"
          },
          "registeredBy" : {
            "type" : "string"
          },
          "registrationDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "size" : {
            "type" : "string"
          },
          "surveySendDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "ticketSendDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "voucher" : {
            "$ref" : "#/components/schemas/Voucher"
          }
        },
        "type" : "object"
      },
...
      "PreSelectionRequest" : {
        "properties" : {
          "comment" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "NONE", "PRE_APPROVED", "PRE_REJECTED", "IN_RESERVE" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Presentation" : {
        "properties" : {
          "accepted" : {
            "type" : "boolean"
          },
          "description" : {
            "minLength" : 1,
            "type" : "string"
          },
          "durationInMinutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "expectedPrice" : {
            "format" : "double",
            "type" : "number"
          },
          "id" : {
            "type" : "string"
          },
          "isFree" : {
            "type" : "boolean"
          },
          "language" : {
            "minLength" : 1,
            "type" : "string"
          },
          "level" : {
            "minLength" : 1,
            "type" : "string"
          },
          "maxGroupSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "new" : {
            "type" : "boolean"
          },
          "preSelectionComment" : {
            "type" : "string"
          },
          "preSelectionStatus" : {
            "enum" : [ "NONE", "PRE_APPROVED", "PRE_REJECTED", "IN_RESERVE" ],
            "type" : "string"
          },
          "ratingEnabled" : {
            "type" : "boolean"
          },
          "ratings" : {
            "items" : {
              "$ref" : "#/components/schemas/Rate"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "shortDescription" : {
            "minLength" : 1,
            "type" : "string"
          },
          "speaker" : {
            "$ref" : "#/components/schemas/Presentation",
            "writeOnly" : true
          },
          "speakers" : {
            "items" : {
              "$ref" : "#/components/schemas/User"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "status" : {
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "$ref" : "#/components/schemas/Tag"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "title" : {
            "minLength" : 1,
            "type" : "string"
          },
          "workshop" : {
            "type" : "boolean"
          }
        },
        "required" : [ "description", "language", "level", "shortDescription", "speakers", "title" ],
        "type" : "object"
      },
...
      "Rate" : {
        "properties" : {
          "comment" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "presentation" : {
            "$ref" : "#/components/schemas/Presentation"
          },
          "reviewerToken" : {
            "type" : "string"
          },
          "value" : {
            "enum" : [ "AWESOME", "GREAT", "IT_WAS_FINE", "BAD", "TERRIBLE" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "RateRequest" : {
        "properties" : {
          "comment" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "reviewerToken" : {
            "type" : "string"
          },
          "value" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "SendMailsRequest" : {
        "properties" : {
          "messageInfoList" : {
            "items" : {
              "$ref" : "#/components/schemas/MessageInfo"
            },
            "type" : "array"
          },
          "template" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Speaker" : {
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Tag" : {
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "User" : {
        "properties" : {
          "admin" : {
            "type" : "boolean"
          },
          "bio" : {
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string"
          },
          "createdDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "github" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "lastModifiedBy" : {
            "type" : "string"
          },
          "lastModifiedDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "origin" : {
            "type" : "string"
          },
          "participant" : {
            "type" : "boolean"
          },
          "participationData" : {
            "$ref" : "#/components/schemas/ParticipationData"
          },
          "photo" : {
            "type" : "string"
          },
          "privacyPolicyAccepted" : {
            "type" : "boolean"
          },
          "socialId" : {
            "type" : "string"
          },
          "speaker" : {
            "type" : "boolean"
          },
          "twitter" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "volunteer" : {
            "type" : "boolean"
          },
          "www" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ViewPresentationRate" : {
        "properties" : {
          "avgRating" : {
            "type" : "number"
          },
          "presentationId" : {
            "type" : "string"
          },
          "ratingCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "ratingCountAwesome" : {
            "format" : "int64",
            "type" : "integer"
          },
          "ratingCountBad" : {
            "format" : "int64",
            "type" : "integer"
          },
          "ratingCountGreat" : {
            "format" : "int64",
            "type" : "integer"
          },
          "ratingCountItWasFine" : {
            "format" : "int64",
            "type" : "integer"
          },
          "ratingCountTerrible" : {
            "format" : "int64",
            "type" : "integer"
          },
          "ratings" : {
            "items" : {
              "$ref" : "#/components/schemas/Rate"
            },
            "type" : "array"
          },
          "speakers" : {
            "items" : {
              "$ref" : "#/components/schemas/Speaker"
            },
            "type" : "array"
          },
          "title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Voucher" : {
        "properties" : {
          "allegro" : {
            "$ref" : "#/components/schemas/AllegroContext"
          },
          "comment" : {
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string"
          },
          "createdDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "emailSent" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string"
          },
          "lastModifiedBy" : {
            "type" : "string"
          },
          "lastModifiedDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "originalBuyer" : {
            "type" : "string"
          },
          "ticketSendDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "type" : {
            "enum" : [ "PARTICIPANT", "SPEAKER", "SPONSOR" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "Jelatyna API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/mailing" : {
      "post" : {
        "operationId" : "sendMails",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SendMailsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "mail-controller" ]
      }
    },
...
    "/presentations/{presentationId}/accept" : {
      "post" : {
        "operationId" : "accept",
        "parameters" : [ {
          "in" : "path",
          "name" : "presentationId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "admin-presentation-controller" ]
      }
    },
...
    "/presentations/{presentationId}/pre-selection" : {
      "post" : {
        "operationId" : "setPreSelection",
        "parameters" : [ {
          "in" : "path",
          "name" : "presentationId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PreSelectionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "admin-presentation-controller" ]
      }
    },
...
    "/presentations/{presentationId}/rating/disable" : {
      "post" : {
        "operationId" : "disableRating",
        "parameters" : [ {
          "in" : "path",
          "name" : "presentationId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "admin-presentation-controller" ]
      }
    },
    "/presentations/{presentationId}/rating/enable" : {
      "post" : {
        "operationId" : "enableRating",
        "parameters" : [ {
          "in" : "path",
          "name" : "presentationId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "admin-presentation-controller" ]
      }
    },
    "/presentations/{presentationId}/ratings" : {
      "get" : {
        "operationId" : "rates_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "presentationId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ViewPresentationRate"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "presentation-controller" ]
      },
      "post" : {
        "operationId" : "addRating",
        "parameters" : [ {
          "in" : "path",
          "name" : "presentationId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "presentation-controller" ]
      }
    },
...
    "/presentations/{presentationId}/reject" : {
      "post" : {
        "operationId" : "reject",
        "parameters" : [ {
          "in" : "path",
          "name" : "presentationId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "admin-presentation-controller" ]
      }
    },
...
    "/ratings" : {
      "post" : {
        "operationId" : "rates",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ViewPresentationRate"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "admin-presentation-controller" ]
      }
    },
...
    "/tags" : {
      "get" : {
        "operationId" : "getAllTags",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Tag"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "presentation-controller" ]
      },
      "post" : {
        "operationId" : "saveTags",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "admin-presentation-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}